update express logic, include new field naming, and edit/delete of dev

+++ app.js
@@ -3,6 +3,7 @@ import bodyParser from 'body-parser';
 import { sendNewDeveloper } from './database/mongodb/models/developer.js';
 import { getDeveloperList } from './database/mongodb/getDevelopers.js';
 import { deleteDeveloper } from './database/mongodb/deleteDeveloper.js';
+import { updateDeveloper } from './database/mongodb/updateDeveloper.js';
 
 const app = express();
 
@@ -18,26 +19,26 @@ app.get('/', (req, res) => {
 });
 
 app.get('/add', (req, res) => {
-    res.render('addDeveloper');
+    res.render('addDeveloper', { developer: {}, verb: 'New', button: 'Submit', func: 'changeDeveloper' });
 });
 
-app.post('/addDeveloper', (req, res) => {
+app.post('/changeDeveloper', (req, res) => {
     let datetime = new Date();
     let lastUpdated = datetime.toLocaleString();
     
     let developer = {
-        name: req.body.developerName,
-        email: req.body.developerEmail,
-        phone: req.body.developerPhone,
-        location: req.body.developerLocation,
-        image: req.body.developerImageUrl,
-        pph: req.body.developerPph,
+        name: req.body.name,
+        email: req.body.email,
+        phone: req.body.phone,
+        location: req.body.location,
+        image: req.body.image,
+        pph: req.body.pph,
         currency: req.body.currency,
-        technology: req.body.developerTechLanguage,
-        description: req.body.developerDescription,
-        yoe: req.body.developerYoe,
-        language: req.body.developerNativeLanguage,
-        linkedin: req.body.developerLinkedIn,
+        technology: req.body.technology,
+        description: req.body.description,
+        yoe: req.body.yoe,
+        language: req.body.language,
+        linkedin: req.body.linkedin,
         lastUpdated: lastUpdated
     };
 
@@ -49,14 +50,25 @@ app.post('/addDeveloper', (req, res) => {
     
 });
 
-app.post('/updateDev', (req, res) => {
-    deleteDeveloper(req.body.email).then(() => {
-        res.render('developerStatus', { devStatus: 'successfully deleted' });
-    }).catch(() => {
-        res.render('developerStatus', { devStatus: 'not deleted. Please try again' });
+app.post('/editDev', (req, res) => {
+    if (req.body.deleteButton) {
+        deleteDeveloper(req.body.email).then(() => {
+            res.render('developerStatus', { devStatus: 'successfully deleted' });
+        }).catch(() => {
+            res.render('developerStatus', { devStatus: 'not deleted. Please try again' });
+        });
+    } else {
+        res.render('changeDeveloper', { developer: req.body, verb: 'Edit', button: 'Update', func: 'updateDeveloper' });
+    }
+});
+
+app.post('/updateDeveloper', (req, res) => {
+    updateDeveloper(req.body.id, req.body).then(() => {
+        res.render('developerStatus', { devStatus: 'successfully updated' });
+    }).catch(error => {
+        res.render('developerStatus', { devStatus: 'not updated. Please try again' });
     });
-    
-})
+});
 
 app.listen('3000', () => {
     console.log('Server is running on port 3000');
